rebase and simplify
