+++ src/component/_generated/api.ts
@@ -12,6 +12,7 @@ import type * as complete from "../complete.js";
 import type * as config from "../config.js";
 import type * as crons from "../crons.js";
 import type * as danger from "../danger.js";
+import type * as errors from "../errors.js";
 import type * as future from "../future.js";
 import type * as kick from "../kick.js";
 import type * as lib from "../lib.js";
@@ -34,6 +35,7 @@ const fullApi: ApiFromModules<{
   config: typeof config;
   crons: typeof crons;
   danger: typeof danger;
+  errors: typeof errors;
   future: typeof future;
   kick: typeof kick;
   lib: typeof lib;
